Extract Read page into seperate components.
- <Reader /> top level component (might not be needed)
- <FeedSidebar />
– <FeedItemPanel />
